+++ src/main/java/model/Register.java
@@ -1,38 +1,27 @@
 package model;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Random;
+
+@AllArgsConstructor
+@Getter
+@Setter
 public class Register {
 
     private String email;
     private String password;
     private String name;
 
-    public Register(String email, String password, String name) {
-        this.email = email;
-        this.password = password;
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
+    public static Register getDefaultRegister() {
+        Random random = new Random();
+        String randomEmail = random.nextInt(10000) + "data-test@example.com";
+        return new Register(
+                randomEmail,
+                "praktikum",
+                "Josh"
+        );
     }
 }
