--- src/main/java/model/RegisterFactory.java
@@ -1,11 +1,0 @@
-package model;
-
-public class RegisterFactory {
-    public static Register getDefaultRegister() {
-        return new Register(
-                "data-test@example.com",
-                "praktikum",
-                "Josh"
-        );
-    }
-}
