+++ src/main/java/model/UserCredentials.java
@@ -1,27 +1,18 @@
 package model;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
 public class UserCredentials {
+
     private String email;
     private String password;
 
-    public UserCredentials(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
+    public static UserCredentials getCredentials(Register register) {
+        return new UserCredentials(register.getEmail(), register.getPassword());
     }
 }
