--- src/main/java/model/UserCredentialsFactory.java
@@ -1,7 +1,0 @@
-package model;
-
-public class UserCredentialsFactory {
-    public static UserCredentials getCredentials(Register register) {
-        return new UserCredentials(register.getEmail(), register.getPassword());
-    }
-}
