--- src/main/java/model/UserFactory.java
@@ -1,8 +1,0 @@
-package model;
-
-public class UserFactory {
-
-    public static User getUser(Register register) {
-        return new User(register.getEmail(), register.getName());
-    }
-}
